requirements/base.txt
# Configuration
django-environ==0.4.5

# Forms
django-braces==1.14.0
django-crispy-forms==1.9.2
django-floppyforms==1.9.0

# Models
django-model-utils==4.0.0

# images
Pillow==7.2.0
sorl-thumbnail==12.6.3

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.42.0

# For the persistance stores
mysqlclient==2.0.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.8

# Useful things
# Time zones support
pytz==2020.1
lxml==4.5.2
cssselect==1.1.0
# Your custom requirements go here
django-guardian==2.3.0
django-mailbox==4.8.1
talon==1.4.4
django-filter==2.3.0
django-autocomplete-light==3.5.0
https://github.com/ad-m/django-flatpages-tinymce/archive/master.zip
django-atom==0.16.3
django-tinymce==3.0.2
django-tinycontent==0.8.0
icalendar==4.0.6
html2text==2020.1.16
pyGravatar==0.0.6
django-bootstrap-pagination==1.7.1
raven==6.10.0
django-import-export==2.3.0
python-dateutil==2.8.1
monotonic==1.5
cached-property==1.5.1
future==0.18.2
django-github-revision==0.0.3
unicodecsv==0.14.1
zipstream-new==1.1.7
django-teryt-tree==0.16.3
django-mptt==0.11.0
markdown-it-py==0.5.5